finish order

+++ frontend/src/screens/SignupScreen.js
@@ -10,23 +10,31 @@ import Button from 'react-bootstrap/Button'
 import { toast } from 'react-toastify';
 import { getError } from '../utils';
 
-function SigninScreen() {
+function SignupScreen() {
   const {search} = useLocation();
   const params = new URLSearchParams(search);
   const redirect = params.get('redirect')?params.get('redirect'):'/'
   const navigate = useNavigate();
   const { state, dispatch: ctxDispatch } = useContext(Store)
   const { userInfo } = state;
+  const [name, setName] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
 
   const submitHandler = async (e) => {
     e.preventDefault();
+    if(password !== confirmPassword) {
+      toast.error('passwords do not match');
+      return;
+    }
     try{
-      const {data} = await axios.post('/api/users/signin', {
+      const {data} = await axios.post('/api/users/signup', {
+        name,
         email,
         password
       })
+      console.log(data)
       ctxDispatch({type: 'SIGNIN_USER', payload: data})
       localStorage.setItem('userInfo', JSON.stringify(data));
       navigate(redirect || '/')
@@ -42,11 +50,15 @@ function SigninScreen() {
   return (
     <Container className='small-container'>
     <Helmet>
-      <title>Sign in</title>
+      <title>Sign up</title>
     </Helmet>
-      <h1>Sign in</h1>
+      <h1>Sign up</h1>
       <Form onSubmit={submitHandler}>
-        <Form.Group className='mb-3' controlId="email">
+        <Form.Group className='mb-3' controlId="name">
+        <Form.Label>Name</Form.Label>
+        <Form.Control type="text" required onChange={(e)=>setName(e.target.value)} />
+      </Form.Group>
+        <Form.Group>
           <Form.Label>Email</Form.Label>
           <Form.Control type="email" required onChange={(e)=>setEmail(e.target.value)} />
         </Form.Group>
@@ -54,16 +66,20 @@ function SigninScreen() {
           <Form.Label>Password</Form.Label>
           <Form.Control type="password" required onChange={(e)=>setPassword(e.target.value)}></Form.Control>
         </Form.Group>
+        <Form.Group className='mb-3' controlId="confirmPassword">
+          <Form.Label>Confirm Password</Form.Label>
+          <Form.Control type="password" required onChange={(e)=>setConfirmPassword(e.target.value)}></Form.Control>
+        </Form.Group>
         <div className="mb-3">
-          <Button type="submit" >Sign In</Button>
+          <Button type="submit" >Sign Up</Button>
         </div>
         <div className="mb-3">
-          New Customer? {'  '}
-          <Link to={`/signup?redirect=${redirect}`}>Create your account</Link>
+          Already have an account? {'  '}
+          <Link to={`/signin?redirect=${redirect}`}>Sign In</Link>
         </div>
       </Form>
     </Container>
   )
 }
 
-export default SigninScreen
+export default SignupScreen
